% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_download_csv.r
\name{osf_download_csv}
\alias{osf_download_csv}
\title{Download all comma-separated-value (\code{.csv}) files from an \acronym{OSF} project or component.}
\usage{
osf_download (
  id,
  path = 'data-raw/'
)
}
\arguments{
\item{id}{An \acronym{OSF} identifier corresponding to an \acronym{OSF} project or component.}

\item{path}{Path pointing to a local directory where the downloaded files will be saved.
Default is to use \file{data-raw/}.}
}
\description{
Download all \code{.csv} files from an existing \acronym{OSF} project or component
based on the associated unique identifier. This is a 5-character global unique identifier (\acronym{GUID}).
See below for details. Note that this function only downloads files listed in the top-level directory.
}
\details{
See package \link[osfr]{osfr} for additional information.
}
\examples{
osf_download_csv('5ad84')
osf_download_csv('5ad84', 'new_data/')
}
