% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfrext-package.R
\docType{package}
\name{osfrext-package}
\alias{osfrext}
\alias{osfrext-package}
\title{osfrext: osfr Extension}
\description{
Functions that streamline common OSF actions in initializing data analysis projects.
}
\details{
This package extends the functionality of the \link[osfr]{osfr} package by
condensing commonly performed tasks into single functions.
}
\author{
\strong{Maintainer}: Dylan Russell \email{dyl.russell@gmail.com} (\href{https://orcid.org/0000-0002-9543-9897}{ORCID})

}
\keyword{internal}
